+++ pi.key.pub
@@ -1,0 +1,1 @@
+ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDjEkEESXUNFryWjGnYWceyo8fdt961Gl8ga3SDmqLLU9Z5vLBsb1gsLnUy9F4h/KDmcYyIb9exzTs/Wqj7ga2bOjodvYAyLXneVEIkcSOc+Skx7A5x6vTOM5qfP0O0eDSQ7Ym2/MoZklKkR+K9nRw84FtFCRy0WZMJj5QOTLwSIB/NeM39gMu9WMIscL1ZiyG60jdaKcZJ64TYkKgii3LZ5ygZCol6fhiupv5N0/oCujAihoL3ocbmRQIx12mSqaPyU33kXBUuXlFqZxbf+DPjDFtYFZ/RZ2bPuhvhOa/SM6qhOVlN8O+rN3MQchQGThksor76H0aRZUWZtFBtOFAT cdowling@Claytons-MacBook-Pro.local
